Solved day 1 part 2
Decided to not enumerate input files and instead created a
'test_solver' method that takes the solver, the input file path,
and an optional expected value.    The result is printed and an
error is thrown if the expected argument is passed and doesn't
match the result returned from the solver.

There are commented print statements I used for debugging.